Print end color escape sequence only when colors are enabled.
Printing "\e[0m" is not a problem for major terminal emulators, but
not for Emacs' dump terminal... Progress bar is just not shown when
fuubar works inside Emacs' terminal, even when colors are disabled.

With this patch a user is able to disable colors and have progress bar
displayed in Emacs.