Bugfix: Fix duplicate constant warnings
From the RC to the final release, RSpec changed their base formatter so that it
would include `rspec/core/formatters/console_codes`.  Due to this, and because
we were requiring it in fuubar as well, users were getting warning about the
fact that constants had already been defined.

Now we're just requiring `rspec/core/formatters/base_text_formatter` and
allowing it to require `console_codes` for us.

Issue URLs:
  * https://github.com/thekompanee/fuubar/issues/69
  * https://github.com/thekompanee/fuubar/issues/70
  * https://github.com/thekompanee/fuubar/pull/72

Actions:
  References #69
  References #70
  References #72

--------------------------------------------------------------------------------