Changed: to use our custom Output object
This is fairly simple.  We're just creating an object which will delegate to the
output object given to use by RSpec for every method except for `tty?`.  In that
case we'll first check to see if we have a TTY override before finally asking
our output object.

--------------------------------------------------------------------------------
My Tracker Issue URLs:
  * https://github.com/thekompanee/fuubar/pull/88
Actions:
  * Closes #88
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>