Stop auto refresh when a Pry session starts
This should fix #108.

The problem with #105 was that the Pry-related code was running as soon
as the `fuubar` gem got loaded, which can be before `Pry` is defined (if
the `pry` gem gets loaded later).

In the changes I propose here,

- the Pry hooks are added later in the process, only when Fuubar starts,
which allows:
  - allows to make sure all the gems in the Gemfile were loaded before
  testing the presence of `Pry`
  - avoiding to add Pry hooks if the Fuubar output is not used (eg. I
usually use Fuubar, but sometimes run `rspec --format documentation`)
- I used :before_session and :after_session hooks, allowing Fuubar to
keep refreshing as long as no Pry session is started

There still is at least one limitation: if the `pry` gem is added to
the Gemfile with the option `require: false`, then required too late
for Fuubar to know, then the hooks will not be added, and auto-refresh
will keep happening when the Pry session opens.